refactor: extract shared view logic into view_core module and add DRY macros
- Create view_core.rs with shared view_with_cache() and build_view_cache()
- Add impl_view_cache_methods! macro for cache enable/disable/clear methods
- Add impl_vault_view_methods! macro for complex view methods (get_cap_groups,
  get_pending_governance_actions, market lookups, snapshots, etc.)
- Apply macros to VaultClient, VaultViewClient, and KeyPoolClient
- Remove ~400 lines of duplicated view method implementations
