+++ client/vault/src/methods.rs
@@ -3,6 +3,218 @@
 //! This macro generates the common vault methods for any client that implements
 //! the required helper methods.
 
+/// Generate view cache management methods for a client.
+///
+/// The client must have these fields:
+/// - `view_cache: RwLock<Option<ViewCache>>`
+#[macro_export]
+macro_rules! impl_view_cache_methods {
+    ($client:ty) => {
+        impl $client {
+            pub fn enable_view_cache(&self, capacity: u32, ttl_seconds: u64) {
+                use $crate::lock_ext::RwLockExt;
+                use std::time::Duration;
+
+                if capacity == 0 {
+                    *self.view_cache.write_recover() = None;
+                    return;
+                }
+
+                let cache = $crate::ViewCache::builder()
+                    .max_capacity(capacity as u64)
+                    .time_to_live(Duration::from_secs(ttl_seconds))
+                    .build();
+
+                *self.view_cache.write_recover() = Some(cache);
+            }
+
+            pub fn disable_view_cache(&self) {
+                use $crate::lock_ext::RwLockExt;
+                *self.view_cache.write_recover() = None;
+            }
+
+            pub async fn clear_view_cache(&self) -> Result<(), $crate::ErrorWrapper> {
+                use $crate::lock_ext::RwLockExt;
+                let cache = { self.view_cache.read_recover().clone() };
+                if let Some(cache) = cache {
+                    cache.invalidate_all();
+                }
+                Ok(())
+            }
+        }
+    };
+}
+
+/// Generate complex vault view methods for a client.
+///
+/// The client must have these helper methods:
+/// - `view<T>(&self, account_id: &NearAccountId, method: &str, args: impl Serialize) -> Result<T>`
+/// - `near_id(&self, id: &AccountId) -> Result<NearAccountId, ErrorWrapper>`
+/// - `vault` field of type `NearAccountId`
+#[macro_export]
+macro_rules! impl_vault_view_methods {
+    ($client:ty) => {
+        #[uniffi::export(async_runtime = "tokio")]
+        impl $client {
+            #[instrument(skip(self))]
+            pub async fn get_cap_groups(&self) -> Result<Vec<$crate::CapGroup>, $crate::ErrorWrapper> {
+                let groups = self
+                    .view::<Vec<(
+                        templar_common::vault::CapGroupId,
+                        templar_common::vault::CapGroupRecord,
+                    )>>(&self.vault, "get_cap_groups", ())
+                    .await
+                    .map_err($crate::ErrorWrapper::from)?;
+
+                Ok(groups
+                    .into_iter()
+                    .map(|(id, rec)| $crate::CapGroup {
+                        id: id.into(),
+                        cap: rec.cap.0.to_string(),
+                        relative_cap: u128::from(rec.relative_cap).to_string(),
+                        principal: rec.principal.to_string(),
+                    })
+                    .collect())
+            }
+
+            #[instrument(skip(self))]
+            pub async fn get_pending_governance_actions(
+                &self,
+            ) -> Result<Vec<$crate::PendingGovernanceAction>, $crate::ErrorWrapper> {
+                let pending = self
+                    .view::<Vec<$crate::PendingValueSerde>>(&self.vault, "get_pending_governance_actions", ())
+                    .await
+                    .map_err($crate::ErrorWrapper::from)?;
+
+                Ok(pending
+                    .into_iter()
+                    .map(|p| $crate::PendingGovernanceAction {
+                        action: p.value.into(),
+                        valid_at_ns: p.valid_at_ns,
+                    })
+                    .collect())
+            }
+
+            #[instrument(skip(self, market))]
+            pub async fn get_market_id_of_account(
+                &self,
+                market: &$crate::AccountId,
+            ) -> Result<Option<$crate::MarketId>, $crate::ErrorWrapper> {
+                let res = self
+                    .view::<Option<U64>>(
+                        &self.vault,
+                        "get_market_id_of_account",
+                        (self.near_id(market)?,),
+                    )
+                    .await
+                    .map_err($crate::ErrorWrapper::from)?;
+
+                let Some(u) = res else {
+                    return Ok(None);
+                };
+
+                let id_u32: u32 =
+                    u.0.try_into()
+                        .map_err(|_| $crate::ErrorWrapper::Wrapped("market id out of u32 range".to_string()))?;
+
+                Ok(Some($crate::MarketId(id_u32)))
+            }
+
+            #[instrument(skip(self, market_id))]
+            pub async fn get_market_account_by_id(
+                &self,
+                market_id: $crate::MarketId,
+            ) -> Result<Option<$crate::AccountId>, $crate::ErrorWrapper> {
+                use near_account_id::AccountId as NearAccountId;
+
+                let res = self
+                    .view::<Option<NearAccountId>>(
+                        &self.vault,
+                        "get_market_account_by_id",
+                        (U64::from(market_id.0 as u64),),
+                    )
+                    .await
+                    .map_err($crate::ErrorWrapper::from)?;
+
+                Ok(res.map(|a| $crate::AccountId::from(a.to_string())))
+            }
+
+            #[instrument(skip(self))]
+            pub async fn list_markets_with_ids(&self) -> Result<Vec<$crate::MarketWithId>, $crate::ErrorWrapper> {
+                use near_account_id::AccountId as NearAccountId;
+
+                let res = self
+                    .view::<Vec<(U64, NearAccountId)>>(&self.vault, "list_markets_with_ids", ())
+                    .await
+                    .map_err($crate::ErrorWrapper::from)?;
+
+                let mapped =
+                    res.into_iter()
+                        .map(|(id, account)| {
+                            let id_u32: u32 = id.0.try_into().map_err(|_| {
+                                $crate::ErrorWrapper::Wrapped("market id out of u32 range".to_string())
+                            })?;
+                            Ok($crate::MarketWithId {
+                                market_id: $crate::MarketId(id_u32),
+                                account: $crate::AccountId::from(account.to_string()),
+                            })
+                        })
+                        .collect::<Result<Vec<_>, $crate::ErrorWrapper>>()?;
+
+                Ok(mapped)
+            }
+
+            #[instrument(skip(self))]
+            pub async fn get_vault_snapshot(&self) -> Result<$crate::VaultSnapshot, $crate::ErrorWrapper> {
+                Ok($crate::VaultSnapshot {
+                    configuration: self.get_configuration().await?,
+                    total_assets: self.get_total_assets().await?,
+                    last_total_assets: self.get_last_total_assets().await?,
+                    idle_balance: self.get_idle_balance().await?,
+                    total_supply: self.get_total_supply().await?,
+                    max_deposit: self.get_max_deposit().await?,
+                    max_single_market_deposit: self.get_max_single_market_deposit().await?,
+                    fee_anchor: self.get_fee_anchor().await?,
+                    fees: self.get_fees().await?,
+                    restrictions: self.get_restrictions().await?,
+                    cap_groups: self.get_cap_groups().await?,
+                    pending_governance_actions: self.get_pending_governance_actions().await?,
+                    withdrawing_op_id: self.get_withdrawing_op_id().await?,
+                    has_pending_market_withdrawal: self.has_pending_market_withdrawal().await?,
+                    current_withdraw_request_id: self.get_current_withdraw_request_id().await?,
+                    queue_tail: self.queue_tail().await?,
+                    next_pending_withdrawal_id: self.peek_next_pending_withdrawal_id().await?,
+                    markets_with_ids: self.list_markets_with_ids().await?,
+                })
+            }
+
+            #[instrument(skip(self, markets))]
+            pub async fn resolve_market_ids(
+                &self,
+                markets: &[$crate::AccountId],
+            ) -> Result<Vec<Option<$crate::MarketId>>, $crate::ErrorWrapper> {
+                let mut out = Vec::with_capacity(markets.len());
+                for market in markets {
+                    out.push(self.get_market_id_of_account(market).await?);
+                }
+                Ok(out)
+            }
+
+            #[instrument(skip(self, market_ids))]
+            pub async fn resolve_market_accounts(
+                &self,
+                market_ids: &[$crate::MarketId],
+            ) -> Result<Vec<Option<$crate::AccountId>>, $crate::ErrorWrapper> {
+                let mut out = Vec::with_capacity(market_ids.len());
+                for id in market_ids {
+                    out.push(self.get_market_account_by_id(*id).await?);
+                }
+                Ok(out)
+            }
+        }
+    };
+}
+
 /// Generate vault view and call methods for a client.
 ///
 /// The client must have these helper methods:
